+++ spotinst/commons/managed_instance_aws_integrations/consts.go
@@ -10,8 +10,10 @@ const (
 	Domains            commons.FieldName = "domains"
 	HostedZoneId       commons.FieldName = "hosted_zone_id"
 	SpotinstAcctID     commons.FieldName = "spotinst_acct_id"
+	RecordSetType      commons.FieldName = "record_set_type"
 	RecordSets         commons.FieldName = "record_sets"
 	UsePublicIP        commons.FieldName = "use_public_ip"
+	UsePublicDNS       commons.FieldName = "use_public_dns"
 	Route53Name        commons.FieldName = "name"
 	// -----------------------------------