+++ spotinst/elastigroup_aws_integrations/fields_spotinst_elastigroup_aws_integrations_route53.go
@@ -1,9 +1,6 @@
 package elastigroup_aws_integrations
 
 import (
-	"errors"
-	"fmt"
-
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/spotinst/spotinst-sdk-go/service/elastigroup/providers/aws"
 	"github.com/spotinst/spotinst-sdk-go/spotinst"
@@ -38,12 +35,22 @@ func SetupRoute53(fieldsMap map[commons.FieldName]*commons.GenericField) {
 									Optional: true,
 								},
 
+								string(RecordSetType): {
+									Type:     schema.TypeString,
+									Optional: true,
+								},
+
 								string(RecordSets): {
 									Type:     schema.TypeSet,
 									Required: true,
 									Elem: &schema.Resource{
 										Schema: map[string]*schema.Schema{
 											string(UsePublicIP): {
+												Type:     schema.TypeBool,
+												Optional: true,
+											},
+
+											string(UsePublicDNS): {
 												Type:     schema.TypeBool,
 												Optional: true,
 											},
@@ -62,18 +69,6 @@ func SetupRoute53(fieldsMap map[commons.FieldName]*commons.GenericField) {
 			},
 		},
 		func(resourceObject interface{}, resourceData *schema.ResourceData, meta interface{}) error {
-			egWrapper := resourceObject.(*commons.ElastigroupWrapper)
-			elastigroup := egWrapper.GetElastigroup()
-			var result []interface{} = nil
-			if elastigroup != nil && elastigroup.Integration != nil && elastigroup.Integration.Route53 != nil {
-				result = flattenRoute53Integration(elastigroup.Integration.Route53)
-			}
-
-			if result != nil {
-				if err := resourceData.Set(string(IntegrationRoute53), result); err != nil {
-					return fmt.Errorf(string(commons.FailureFieldReadPattern), string(IntegrationRoute53), err)
-				}
-			}
 			return nil
 		},
 
@@ -138,11 +133,11 @@ func expandAWSGroupRoute53IntegrationDomains(data interface{}) ([]*aws.Domain, e
 
 	for _, v := range list {
 		attr, ok := v.(map[string]interface{})
-		domain := &aws.Domain{}
-
 		if !ok {
 			continue
 		}
+
+		domain := new(aws.Domain)
 
 		if v, ok := attr[string(HostedZoneId)].(string); ok && v != "" {
 			domain.SetHostedZoneID(spotinst.String(v))
@@ -152,6 +147,10 @@ func expandAWSGroupRoute53IntegrationDomains(data interface{}) ([]*aws.Domain, e
 			domain.SetSpotinstAccountID(spotinst.String(v))
 		}
 
+		if v, ok := attr[string(RecordSetType)].(string); ok && v != "" {
+			domain.SetRecordSetType(spotinst.String(v))
+		}
+
 		if r, ok := attr[string(RecordSets)]; ok {
 			if recordSets, err := expandAWSGroupRoute53IntegrationDomainsRecordSets(r); err != nil {
 				return nil, err
@@ -159,8 +158,10 @@ func expandAWSGroupRoute53IntegrationDomains(data interface{}) ([]*aws.Domain, e
 				domain.SetRecordSets(recordSets)
 			}
 		}
+
 		domains = append(domains, domain)
 	}
+
 	return domains, nil
 }
 
@@ -170,63 +171,26 @@ func expandAWSGroupRoute53IntegrationDomainsRecordSets(data interface{}) ([]*aws
 
 	for _, v := range list {
 		attr, ok := v.(map[string]interface{})
-
 		if !ok {
 			continue
 		}
 
-		if _, ok := attr[string(UsePublicIP)]; !ok {
-			return nil, errors.New("invalid record set attributes: use_public_ip missing")
+		recordSet := new(aws.RecordSet)
+
+		if v, ok := attr[string(Name)].(string); ok {
+			recordSet.SetName(spotinst.String(v))
 		}
 
-		if _, ok := attr[string(Name)]; !ok {
-			return nil, errors.New("invalid record set attributes: name missing")
+		if v, ok := attr[string(UsePublicIP)].(bool); ok && v {
+			recordSet.SetUsePublicIP(spotinst.Bool(v))
 		}
 
-		recordSet := &aws.RecordSet{
-			UsePublicIP: spotinst.Bool(attr[string(UsePublicIP)].(bool)),
-			Name:        spotinst.String(attr[string(Name)].(string)),
+		if v, ok := attr[string(UsePublicDNS)].(bool); ok && v {
+			recordSet.SetUsePublicDNS(spotinst.Bool(v))
 		}
 
 		recordSets = append(recordSets, recordSet)
 	}
+
 	return recordSets, nil
-}
-
-func flattenRoute53Integration(route53 *aws.Route53Integration) []interface{} {
-	result := make(map[string]interface{})
-
-	if route53.Domains != nil {
-		result[string(Domains)] = flattenDomain(route53.Domains)
-	}
-
-	return []interface{}{result}
-}
-
-func flattenDomain(domains []*aws.Domain) []interface{} {
-	result := make([]interface{}, 0, len(domains))
-	for _, domain := range domains {
-		m := make(map[string]interface{})
-		m[string(HostedZoneId)] = spotinst.StringValue(domain.HostedZoneID)
-		m[string(SpotinstAcctID)] = spotinst.StringValue(domain.SpotinstAccountID)
-
-		if domain.RecordSets != nil {
-			m[string(RecordSets)] = flattenRecordsSets(domain.RecordSets)
-		}
-
-		result = append(result, m)
-	}
-	return result
-}
-
-func flattenRecordsSets(recordSets []*aws.RecordSet) []interface{} {
-	result := make([]interface{}, 0, len(recordSets))
-	for _, recordSet := range recordSets {
-		m := make(map[string]interface{})
-		m[string(UsePublicIP)] = spotinst.BoolValue(recordSet.UsePublicIP)
-		m[string(Name)] = spotinst.StringValue(recordSet.Name)
-
-		result = append(result, m)
-	}
-	return result
 }